moved all templated functions, and a few non-templated ones, out of log_recovery.c and into log_recovery_redo_log_rec.hpp header such that they are visible to templated class redo_log_rec_entry_templ 'do_work' function; added derived templated class redo_log_rec_entry_templ which is the actual 'workhorse' that the parallel mechanism will use to invoke logic
